+++ devtools/replidev/src/command/mod.rs
@@ -1,3 +1,4 @@
 pub mod certs;
 pub mod conf;
 pub mod deps;
+pub mod play;